requirements-dev.txt
pre-commit

# static analysis
black==24.3.0
flake8==6.0.0
flake8-print
flake8-todo

# safety checking
bandit

# type checking
mypy
sqlalchemy-stubs
types-requests

# tests
pytest

# strawberry.py tools
git+https://github.com/strawberry-py/strawberry-tools.git
